get-started-demo/requirements.txt
azure-storage-blob
nvidia-tao
pycocotools
absl-py
tensorflow-object-detection-api